+++ R/types.R
@@ -67,8 +67,17 @@ is_complex <- function(x, n = NULL, finite = NULL) {
 }
 #' @export
 #' @rdname type-predicates
-is_character <- function(x, n = NULL) {
-  .Call(ffi_is_character, x, n)
+#' @inheritParams args_dots_empty
+#' @param missing,empty A logical value indicating whether to forbid
+#'   (`FALSE`) or enforce (`TRUE`) missing values or empty `""`. If
+#'   `NULL`, these values are ignored.
+is_character <- function(x,
+                         n = NULL,
+                         ...,
+                         missing = NULL,
+                         empty = NULL) {
+  check_dots_empty0(...)
+  .Call(ffi_is_character, x, n, missing, empty)
 }
 #' @export
 #' @rdname type-predicates
@@ -154,11 +163,11 @@ is_scalar_raw <- function(x) {
 #' @export
 #' @param string A string to compare to `x`. If a character vector,
 #'   returns `TRUE` if at least one element is equal to `x`.
-#' @param empty If `NULL`, the empty string `""` is not treated
-#'   specially. If `TRUE`, `x` must be `""`. If `FALSE`, `x` can't be
-#'   `""`. `empty` and `string` are mutually exclusive arguments.
+#' @inheritParams type-predicates
+#' @inheritParams args_dots_empty
 #' @rdname scalar-type-predicates
-is_string <- function(x, string = NULL, empty = NULL) {
+is_string <- function(x, string = NULL, ..., empty = NULL) {
+  check_dots_empty0(...)
   .Call(ffi_is_string, x, string, empty)
 }
 #' @export